+++ cms/src/globals/Footer.ts
@@ -2,6 +2,7 @@ import type { GlobalConfig } from 'payload'
 
 import { isAdminOrEditor } from '../access/isAdminOrEditor'
 import { triggerDeployGlobalAfterChange } from '../hooks/triggerDeploy'
+import { navLinkFields } from '../shared'
 
 export const Footer: GlobalConfig = {
   slug: 'footer',
@@ -26,20 +27,7 @@ export const Footer: GlobalConfig = {
         {
           name: 'links',
           type: 'array',
-          fields: [
-            {
-              name: 'page',
-              type: 'relationship',
-              relationTo: ['pages', 'posts', 'authors'],
-              required: true,
-            },
-            {
-              name: 'label',
-              type: 'text',
-              required: true,
-              localized: true,
-            },
-          ],
+          fields: navLinkFields,
         },
       ],
     },